fix(ws): WebSocketSecurityConfig を2クラスに分割して ChannelSecurityInterceptor エラーを修正
@Bean から匿名サブクラスを返す実装では Spring が ChannelSecurityInterceptor を
登録できない。@Configuration クラス自体が AbstractSecurityWebSocketMessageBrokerConfigurer
を extend する必要があるため、local / prod 用に別クラスに分割。
