+++ mapoker-backend/src/main/java/com/mapoker/infrastructure/config/WebSocketSecurityConfigProd.java
@@ -1,0 +1,24 @@
+package com.mapoker.infrastructure.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.security.config.annotation.web.messaging.MessageSecurityMetadataSourceRegistry;
+import org.springframework.security.config.annotation.web.socket.AbstractSecurityWebSocketMessageBrokerConfigurer;
+
+@Configuration
+@Profile("!local")
+public class WebSocketSecurityConfigProd extends AbstractSecurityWebSocketMessageBrokerConfigurer {
+
+    @Override
+    protected void configureInbound(MessageSecurityMetadataSourceRegistry messages) {
+        messages
+                .nullDestMatcher().authenticated()
+                .simpSubscribeDestMatchers("/topic/**", "/user/queue/**").authenticated()
+                .anyMessage().authenticated();
+    }
+
+    @Override
+    protected boolean sameOriginDisabled() {
+        return true;
+    }
+}
